move configs to src/dr_plotter/configs/
