+++ src/dr_plotter/configs/__init__.py
@@ -1,0 +1,7 @@
+from dr_plotter.configs.legend_config import LegendConfig, LegendStrategy
+from dr_plotter.configs.figure_config import FigureConfig
+from dr_plotter.configs.cycle_config import CycleConfig
+from dr_plotter.configs.grouping_config import GroupingConfig
+from dr_plotter.configs.positioning_config import PositioningConfig
+from dr_plotter.configs.faceting_config import FacetingConfig
+from dr_plotter.configs.plot_config import LayoutConfig, StyleConfig, PlotConfig
