+++ src/dr_plotter/configs/grouping_config.py
