+++ src/dr_plotter/plotters/style_engine.py
@@ -1,8 +1,8 @@
 from typing import Any, Dict, List, Optional
 
 from dr_plotter.channel_metadata import ChannelRegistry
-from dr_plotter.cycle_config import CycleConfig
-from dr_plotter.grouping_config import GroupingConfig
+from dr_plotter.configs.cycle_config import CycleConfig
+from dr_plotter.configs.grouping_config import GroupingConfig
 from dr_plotter.theme import Theme
